Move test class to separate feature folder
